requirements.txt
pymongo[srv]==4.2.0
python-dotenv==0.21.0
from-root==1.0.2
pandas==1.5.1
scikit-learn==1.1.3
PyYAML==6.0
black==22.10.0